% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notes_format.R
\name{atopify}
\alias{atopify}
\title{atopify helper function for notes formatting}
\usage{
atopify(A, i)
}
\arguments{
\item{A}{Vector to atopify}

\item{i}{Recursion index}
}
\value{
atopified caption vector
}
\description{
atopify helper function for notes formatting
}
