% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{clean_palette}
\alias{clean_palette}
\title{Clean color palette}
\usage{
clean_palette(
  n,
  colors = clean_colors,
  type = c("discrete", "continuous"),
  order = NULL
)
}
\arguments{
\item{n}{Number of colors}

\item{colors}{Named list of colors}

\item{type}{Discrete or continuous scale}

\item{order}{Colors to move to the front}
}
\value{
Clean palette object
}
\description{
Clean color palette
}
