% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{colid}
\alias{colid}
\title{Add letters to column names}
\usage{
colid(DT, delim = "/")
}
\arguments{
\item{DT}{table to add column letters}

\item{delim}{delimiter}
}
\value{
table with changed names
}
\description{
Add letters to column names
}
