% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{facetseq}
\alias{facetseq}
\title{facetseq}
\usage{
facetseq(x, m, sq1, sq2)
}
\arguments{
\item{x}{number}

\item{m}{max test}

\item{sq1}{seq 1}

\item{sq2}{seq 2}
}
\value{
sequence
}
\description{
facetseq
}
\examples{
facetseq(c(.1,.5,.9), 1, 0:4, c("0","0.25","0.5", "0.75", "1"))
}
