% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{formatcols}
\alias{formatcols}
\title{format columns}
\usage{
formatcols(
  DT,
  comma_cols = NULL,
  dollar_cols = NULL,
  percent_cols = NULL,
  comma_accuracy = 1,
  dollar_accuracy = 1,
  percent_accuracy = 0.01
)
}
\arguments{
\item{DT}{import table}

\item{comma_cols}{comma columns}

\item{dollar_cols}{dollar columns}

\item{percent_cols}{percent columns}

\item{comma_accuracy}{comma accuracy}

\item{dollar_accuracy}{dollar accuracy}

\item{percent_accuracy}{percent accuracy}
}
\description{
format columns
}
