% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{intoverlapdays}
\alias{intoverlapdays}
\title{intoverlapdays}
\usage{
intoverlapdays(start1, end1, start2, end2)
}
\arguments{
\item{start1}{start date 1}

\item{end1}{end date 1}

\item{start2}{start date 2}

\item{end2}{end date 2}
}
\value{
overlap day number
}
\description{
intoverlapdays
}
