% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{limroundup}
\alias{limroundup}
\title{limroundup}
\usage{
limroundup(min, x, v)
}
\arguments{
\item{min}{vector min}

\item{x}{number}

\item{v}{round to}
}
\value{
limits
}
\description{
limroundup
}
\examples{
limroundup(0, 272, 50)
}
