% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{monthsl}
\alias{monthsl}
\title{monthsl}
\usage{
monthsl(DF, b = 4, f = "\%b-\%y", last = TRUE)
}
\arguments{
\item{DF}{data.table}

\item{b}{buckets}

\item{f}{date format}

\item{last}{include last month}
}
\value{
breaks and labels
}
\description{
monthsl
}
