% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notes_format.R
\name{notes_format}
\alias{notes_format}
\title{Clean caption element of ggplot}
\usage{
notes_format(caption, caption_title = "Notes & Sources:")
}
\arguments{
\item{caption}{Vector of caption lines}

\item{caption_title}{Captiin title}
}
\value{
Formatted caption
}
\description{
Clean caption element of ggplot
}
