% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{scale_color_clean}
\alias{scale_color_clean}
\title{Clean manual color scale for ggplots}
\usage{
scale_color_clean(colors)
}
\arguments{
\item{colors}{Override colors}
}
\value{
Clean manual color scale
}
\description{
Clean manual color scale for ggplots
}
