% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{scale_fill_clean}
\alias{scale_fill_clean}
\title{Clean manual fill scale for ggplots}
\usage{
scale_fill_clean(colors)
}
\arguments{
\item{colors}{Override colors}
}
\value{
Clean manual fill scale
}
\description{
Clean manual fill scale for ggplots
}
