% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{seqroundup}
\alias{seqroundup}
\title{seqroundup}
\usage{
seqroundup(x, v, s)
}
\arguments{
\item{x}{number}

\item{v}{round to}

\item{s}{steps}
}
\value{
breaks
}
\description{
seqroundup
}
\examples{
seqroundup(272, 50, 5)
}
