% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{tab}
\alias{tab}
\title{Format table}
\usage{
tab(
  gt,
  dark = FALSE,
  title = "",
  subtitle = "",
  notes = c(),
  table.font.size = 11,
  title.font.size = 18,
  col.height = 5,
  title.padding = 50
)
}
\arguments{
\item{gt}{table to format}

\item{dark}{dark column labels}

\item{title}{table title}

\item{subtitle}{table subtitle}

\item{notes}{table notes vector}

\item{table.font.size}{table font size}

\item{title.font.size}{title font size}

\item{col.height}{column label row height}

\item{title.padding}{title padding}
}
\value{
html table
}
\description{
Format table
}
