% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{tabpdf}
\alias{tabpdf}
\title{Save formatted html table}
\usage{
tabpdf(
  tab,
  path,
  options = list(landscape = TRUE, marginTop = 0.5, marginBottom = 0.5, marginLeft = 0.5,
    marginRight = 0.5)
)
}
\arguments{
\item{tab}{table to save to html and pdf}

\item{path}{path name}

\item{options}{chrome_print options}
}
\value{
html and pdf
}
\description{
Save formatted html table
}
