% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_clean.R
\name{theme_clean}
\alias{theme_clean}
\title{Clean a ggplot}
\usage{
theme_clean(
  grid = FALSE,
  x_lines = FALSE,
  y_lines = FALSE,
  legend_position = "bottom",
  plot_margin_in = 0.5,
  caption_margin = 0.25,
  title_margin = 0,
  subtitle_margin = 0.25,
  borderizer = FALSE,
  font = "sans",
  text_size = 14,
  title_size = 18,
  subtitle_size = 16,
  caption_size = 11,
  legend_title_size = 14,
  legend_title_face = "plain",
  text_color = "black",
  grid_color = "#D3D3D3",
  no_axis = FALSE,
  axis_label_face = "plain",
  axis_line_width = 0.5,
  grid_line_width = 0.1,
  x_axis_label_angle = 0
)
}
\arguments{
\item{grid}{Add x- and y-axis grid lines}

\item{x_lines}{Add x-axis vertical grid lines}

\item{y_lines}{Add y-axis horizontal grid lines}

\item{legend_position}{Change legend position}

\item{plot_margin_in}{Plot margin (inches)}

\item{caption_margin}{Caption margin (inches)}

\item{title_margin}{Title margin (inches)}

\item{subtitle_margin}{Subtitle margin (inches)}

\item{borderizer}{Margins so plot fits in LaTeX borderizer}

\item{font}{Font name}

\item{text_size}{Default text size for axes (points)}

\item{title_size}{Title size (points)}

\item{subtitle_size}{Subtitle size (points)}

\item{caption_size}{Subtitle size (points)}

\item{legend_title_size}{Legend title size (points)}

\item{legend_title_face}{Legend title font face ("plain", "bold", "italic", or "bold.italic")}

\item{text_color}{Text color}

\item{grid_color}{Grid line color}

\item{no_axis}{Remove axis lines and ticks}

\item{axis_label_face}{Axis label font face ("plain", "bold", "italic", or "bold.italic")}

\item{axis_line_width}{Axis line width}

\item{grid_line_width}{Grid line width}

\item{x_axis_label_angle}{X-axis label angle}
}
\value{
Theme function
}
\description{
Clean a ggplot
}
