% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_epsilon.R
\name{theme_epsilon}
\alias{theme_epsilon}
\title{Clean a ggplot}
\usage{
theme_epsilon(
  grid = FALSE,
  x_lines = FALSE,
  y_lines = FALSE,
  legend_position = "bottom",
  plot_margin_in = 0.5,
  plot_margin_top_in = 0.5,
  caption_margin = 0.25,
  title_margin = 0,
  subtitle_margin = 0.25,
  borderizer = FALSE,
  font = "sans",
  text_size = 14,
  title_size = 18,
  subtitle_size = 16,
  caption_size = 11,
  legend_title_size = 14,
  legend_title_face = "plain",
  text_color = "black",
  grid_color = "#D3D3D3",
  no_axis = FALSE,
  axis_label_face = "plain",
  axis_line_width = 0.5,
  grid_line_width = 0.1,
  x_axis_label_angle = 0
)
}
\arguments{
\item{x_axis_label_angle}{}
}
\description{
Clean a ggplot
}
