% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{tabtex}
\alias{tabtex}
\title{Format LaTeX table}
\usage{
tabtex(
  DT,
  shading = FALSE,
  align = NULL,
  cwidth = "1in",
  cletters = NULL,
  rowspace = NULL,
  boldrow = NULL,
  longtable = FALSE,
  continued = NULL,
  ...
)
}
\arguments{
\item{DT}{data.table}

\item{shading}{add alternate row shading}

\item{align}{latex table align codes}

\item{cwidth}{set column widths}

\item{cletters}{add column sub-header labels, use option "default" for \link{A}, \link{B}, etc.}

\item{rowspace}{add extra blank row after row number(s)}

\item{longtable}{set longtable formatting}

\item{continued}{add continued text to multi-page longtables}

\item{...}{xtable print options}
}
\value{
latex table from data.table
}
\description{
Format LaTeX table
}
