% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{tabtex}
\alias{tabtex}
\title{Format LaTeX table}
\usage{
tabtex(
  DT,
  shading = TRUE,
  cwidth = "1in",
  cletters = paste0("[", LETTERS[1:ncol(DT)], "]"),
  rowspace = NULL,
  ...
)
}
\arguments{
\item{DT}{data.table}

\item{shading}{add alternate row shading}

\item{cwidth}{set column widths}

\item{cletters}{add column sub-header labels}

\item{rowspace}{add extra space after rows}

\item{...}{xtable print options}
}
\value{
latex table from data.table
}
\description{
Format LaTeX table
}
