% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_clean.R
\name{theme_clean}
\alias{theme_clean}
\title{Clean a ggplot}
\usage{
theme_clean(
  p,
  grid = FALSE,
  xlines = FALSE,
  ylines = FALSE,
  legend_position = "bottom",
  plot_margin_in = 0.5,
  caption = c(),
  caption_title = "Notes & Sources:",
  caption_margin = 0,
  title_t_margin = 0,
  title_b_margin = 0,
  subtitle_t_margin = 0.1,
  subtitle_b_margin = 0.25,
  borderizer = FALSE,
  font = "sans",
  text_size = 14,
  title_size = 18,
  subtitle_size = 16,
  caption_size = 11,
  strip_title_size = 14,
  legend_title_size = 11,
  legend_item_size = 14,
  legend_title_face = "bold",
  legend_rows = NULL,
  text_color = "black",
  majorgrid_color = "#a5a5a5",
  minorgrid_color = "#D3D3D3",
  no_axis = FALSE,
  axis_label_face = "plain",
  axis_line_width = 0.5,
  grid_line_width = 0.1,
  x_axis_label_size = 14,
  x_axis_label_angle = 0,
  x_axis_label_hjust = NULL,
  x_axis_label_vjust = NULL,
  y_axis_label_size = 14,
  y_axis_label_angle = 0,
  lims = NULL,
  breaks = NULL,
  labels = NULL,
  expand = NULL,
  xlims = NULL,
  ylims = NULL,
  xbreaks = waiver(),
  xminorbreaks = waiver(),
  ybreaks = waiver(),
  yminorbreaks = waiver(),
  xlabels = waiver(),
  ylabels = waiver(),
  xexpand = c(0, 0.05),
  yexpand = c(0, 0.05),
  ...,
  save_filename = NULL,
  save_paper_size = "ledger",
  save_orientation = "landscape",
  save_width = NULL,
  save_height = NULL
)
}
\arguments{
\item{p}{ggplot object}

\item{grid}{Add x- and y-axis grid lines}

\item{xlines}{Add x-axis vertical grid lines}

\item{ylines}{Add y-axis horizontal grid lines}

\item{legend_position}{Change legend position}

\item{plot_margin_in}{Plot margin (inches)}

\item{caption}{Caption as a vector, each element appears on new line}

\item{caption_title}{Caption title}

\item{caption_margin}{Caption margin (inches)}

\item{title_t_margin}{Title top margin (inches)}

\item{title_b_margin}{Title bottom margin (inches)}

\item{subtitle_t_margin}{Subtitle top margin (inches)}

\item{subtitle_b_margin}{Subtitle bottom margin (inches)}

\item{borderizer}{Margins so plot fits in LaTeX borderizer}

\item{font}{Font name}

\item{text_size}{Default text size for axes (points)}

\item{title_size}{Title size (points)}

\item{subtitle_size}{Subtitle size (points)}

\item{caption_size}{Subtitle size (points)}

\item{strip_title_size}{Strip title size}

\item{legend_title_size}{Legend title size (points)}

\item{legend_item_size}{Legend item size}

\item{legend_title_face}{Legend title font face}

\item{legend_rows}{Number of legend rows}

\item{text_color}{Text color}

\item{majorgrid_color}{Grid line color}

\item{minorgrid_color}{Grid line color}

\item{no_axis}{Remove axis lines and ticks}

\item{axis_label_face}{Axis label font face}

\item{axis_line_width}{Axis line width}

\item{grid_line_width}{Grid line width}

\item{x_axis_label_size}{x-axis label size}

\item{x_axis_label_angle}{x-axis label angle}

\item{x_axis_label_hjust}{x-axis label hjust}

\item{x_axis_label_vjust}{x-axis label vjust}

\item{y_axis_label_size}{y-axis label size}

\item{y_axis_label_angle}{y-axis label angle}

\item{lims}{Set x and y axis limits}

\item{breaks}{Set x and y axis breaks}

\item{labels}{Set x and y axis labels}

\item{expand}{Set x and y axis expansion percentage}

\item{xlims}{Set x limits}

\item{ylims}{Set y limits}

\item{xbreaks}{Set x breaks}

\item{xminorbreaks}{Set x minor breaks}

\item{ybreaks}{Set y breaks}

\item{yminorbreaks}{Set y minor breaks}

\item{xlabels}{Set x labels}

\item{ylabels}{Set y labels}

\item{xexpand}{Set x expansion}

\item{yexpand}{Set y expansion}

\item{save_filename}{Optional filename to save plot}

\item{save_paper_size}{Paper size of saved plot}

\item{save_orientation}{Orientation of saved plot}

\item{save_width}{Manually set paper width}

\item{save_height}{Manually set paper height}
}
\value{
Theme function
}
\description{
Clean a ggplot
}
